feat(modularize): separate out response, error, and status
