+++ src/error.rs
@@ -1,0 +1,5 @@
+use hyper;
+
+pub struct Error {
+    pub inner: hyper::Error,
+}
