+++ src/status.rs
@@ -1,0 +1,4 @@
+pub enum Status {
+    Ok,
+    NotFound,
+}
